refactor: App클래스 파일 위치 변경
기존 controller 폴더에 위치해있던 App클래스에 대한 역할을 Game클래스로 위임하였기에, App클래스는 파일 위치를 상위 레벨로 위치시켰습니다.
