+++ src/index.js
@@ -1,4 +1,4 @@
-import App from './controller/App.js';
+import App from './App.js';
 
 const app = new App();
 await app.run();
